backend/requirements.txt

fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
chromadb
llama-cpp-python
PyMuPDF
python-docx

